modified auth to check for cookies, changed data type for stock table to decimal and changed user_stocks to reflect these changes

+++ backend/app/models/users_table.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String,Column ,Float, Integer,Date
+from sqlalchemy import String,Column ,Float, Integer,Date, Numeric
 from sqlalchemy import ForeignKey
 from ..db import Base 
 from sqlalchemy.orm import relationship
@@ -13,7 +13,7 @@ class Users(Base):
     hashed_password=Column(String)
     email=Column(String)
     full_name=Column(String)
-    total_capital=Column(Float,default=10000)
+    total_capital=Column(Numeric(10,2),default=10000)
     holdings=relationship("UserStocks",back_populates="user")
 
 
@@ -25,8 +25,8 @@ class UserStocks(Base):
     username=Column(String,ForeignKey("users.username"))
     date=Column(Date)
     stock=Column(String)
-    shares=Column(Float)
-    avg_buy_price:Optional[float]=Column(Float) # type: ignore
+    shares=Column(Numeric(10,2))
+    avg_buy_price:Optional[float]=Column(Numeric(10,2)) # type: ignore
 
     user=relationship("Users", back_populates="holdings")
 
@@ -44,7 +44,7 @@ class Transanctions(Base):
     username=Column(String, ForeignKey("users.username"))
     date=Column(Date)
     stock=Column(String)
-    shares=Column(Float)
+    shares=Column(Numeric(10,2))
     activity = Column(DbEnum(Activity), nullable=False)
